Rewrite page/index logic
This way a handler isn't added for each page, instead the generic
handler looks up if the path matches a page, and serves that. This way I
can i.e. add pages dynamically when the server is running in the future.

A mutex had to be added to the index because I got a concurrent map
writes error, I guess because pages are added in parallell in a
goroutine.